Fixed multiple issues with constraints
The global matrix was wrong for several reasons.
Firstly, python treats 0 as false, so springs to/from the zeroth point were being counted as fixed.
Secondly, when adding constraint sets, it would add sets of triangles.
Two triangles on a set of four points had an overlapping diagonal, which meant two constraints were added, doubling the value in the global matrix.

Also fixed several other issues. It currently is in a semi-working flag-like state.